Add files via upload
A prime number (or a prime) is a natural number greater than 1 that cannot be formed by multiplying two
smaller natural numbers.
“Happy” has many definitions (even in mathematics). For our purposes, a happy number is a number
defined by the following process: Starting with any positive integer, replace the number by the sum of the
squares of its digits, and repeat the process until the number either equals 1 (where it will stay), or it loops
endlessly in a cycle that does not include 1. Those numbers for which this process ends in 1 are happy
numbers, while those that do not end in 1 are unhappy numbers (or sad numbers). For example, 19 is
happy, as the associated sequence is:
1
2 + 92 = 82
8
2 + 22 = 68
6
2 + 82 = 100
1
2 + 02 + 02 = 1
Write code that endless loops, requesting/accepting an int from the user. Given the input, your code
should print out whether it is a happy prime, sad prime, happy non-prime, or sad non-prime.